+++ application/controllers/Engineers.php
@@ -33,7 +33,58 @@ class Engineers extends MY_Controller{
                 $result[] = file_exists('./uploads/'.$eng->photo) ? "<img src='".base_url('uploads/'.$eng->photo)."' width='30px' />" : "-";
                 $result[] = status_switch($eng->status,"engineers/status/$eng->ID/$eng->status");
                 $result[] = "<button type=\"button\" class=\"btn btn-info btn-sm waves-effect waves-light\" onclick='download(\"engineers\",\"".$eng->ID."\")'><i class=\"ti-download\"></i> Download</button>";
-                $result[] = view_button(base_url('engineers/single_engineer/'.$eng->ID))." ".edit_button(base_url('engineers/single_engineer/'.$eng->ID))." ".delete_button(base_url('engineers/delete/'.$eng->ID));
+                $result[] = pass_change_button(base_url('engineers/change_password/'.$eng->ID))." ".view_button(base_url('engineers/single_engineer/'.$eng->ID))." ".edit_button(base_url('engineers/single_engineer/'.$eng->ID))." ".delete_button(base_url('engineers/delete/'.$eng->ID));
+                $table['data'][] = $result;
+            }
+            echo json_encode($table);
+        }else{
+            echo '{
+				"sEcho": 1,
+				"iTotalRecords": "0",
+				"iTotalDisplayRecords": "0",
+				"aaData": []
+			}';
+        }
+    }
+
+    public function working_history($ID){
+        $this->cm->table_name = "sites";
+        $this->cm->where = array('engineerID'=>$ID);
+        $sites = $this->cm->get();
+        if($sites->num_rows() > 0){
+            $n=1;
+            foreach($sites->result() as $site){
+                $result = array();
+                $result[] = $n++;
+                $result[] = $site->name;
+                $result[] = $site->address;
+                $result[] = status_button($site->status);
+                $table['data'][] = $result;
+            }
+            echo json_encode($table);
+        }else{
+            echo '{
+				"sEcho": 1,
+				"iTotalRecords": "0",
+				"iTotalDisplayRecords": "0",
+				"aaData": []
+			}';
+        }
+    }
+
+    public function salary_table($ID){
+        $this->cm->table_name = "engineersalary";
+        $this->cm->where = array('engineerID'=>$ID);
+        $salary = $this->cm->get();
+        if($salary->num_rows() > 0){
+            $n=1;
+            foreach($salary->result() as $sal){
+                $result = array();
+                $result[] = $n++;
+                $result[] = $sal->salary." ৳";
+                $result[] = payment_status($sal->paid,base_url('salary/give_salary_to_engineer/'.$sal->ID.'/'.$sal->paid));
+                $result[] = date('F Y',strtotime($sal->date));
+                $result[] = delete_button(base_url('salary/delete_engineer_salary/'.$sal->ID));
                 $table['data'][] = $result;
             }
             echo json_encode($table);
@@ -67,7 +118,6 @@ class Engineers extends MY_Controller{
                     $upload = $this->cm->upload($key,'./uploads/');
                     if($upload['status'] == 'success'){
                         $filenames[$key] = $upload['file_name'];
-                        $name[$key]['filename'] = $_FILES[$key]['name'];
                     }else{
                         $filenames[$key] = $upload['error'];
                     }
@@ -80,9 +130,9 @@ class Engineers extends MY_Controller{
                 'address' => $post_data['address'],
                 'join_date' => $post_data['join_date'],
                 'salary' => $post_data['salary'],
-                'photo' => $filenames['photo'],
-                'filename' => $name['document']['filename'],
-                'document' => $filenames['document'],
+                'photo' => isset($filenames['photo']) ? $filenames['photo'] : "No File Selected",
+                'filename' => $_FILES['document']['name'] ? $_FILES['document']['name'] : "No File Selected",
+                'document' => isset($filenames['document']) ? $filenames['document'] : "No File Selected",
                 'username' => $post_data['username'],
                 'password' => $this->hash($post_data['password']),
             );
@@ -141,16 +191,12 @@ class Engineers extends MY_Controller{
             $this->send_warning(validation_errors());
         }else{
             $post_data = $this->array_from_post(array('name','email','phone','address','join_date','salary','username'));
-            $filenames['photo'] = $eng->photo;
-            $filenames['document'] = $eng->document;
-            $name['document']['filename'] = $eng->filename;
             if($_FILES){
                 foreach($_FILES as $key => $value){
                     $upload = $this->cm->upload($key,'./uploads/');
                     if($upload['status'] == 'success'){
-                        $this->cm->delete_file('./uploads/'.$filenames[$key]);
+                        $this->cm->delete_file('./uploads/'.$eng->$key);
                         $filenames[$key] = $upload['file_name'];
-                        $name[$key]['filename'] = $_FILES[$key]['name'];
                     }
                 }
             }
@@ -161,9 +207,9 @@ class Engineers extends MY_Controller{
                 'address' => $post_data['address'],
                 'join_date' => $post_data['join_date'],
                 'salary' => $post_data['salary'],
-                'photo' => $filenames['photo'],
-                'filename' => $name['document']['filename'],
-                'document' => $filenames['document'],
+                'photo' => isset($filenames['photo']) ? $filenames['photo'] : $eng->photo,
+                'filename' => $_FILES['document']['name'] ? $_FILES['document']['name'] : $eng->filename,
+                'document' => isset($filenames['document']) ? $filenames['document'] : $eng->document,
                 'username' => $post_data['username'],
             );
             $this->cm->table_name = "engineers";
@@ -173,6 +219,23 @@ class Engineers extends MY_Controller{
                 $this->send_success('Engineer Updated Successfully');
             }else{
                 $this->send_error('Can\' Upudate Engineer Now, Please Try Again');
+            }
+        }
+    }
+
+    public function change_password($ID=NULL){
+        $this->form_validation->set_rules('password','Password','trim|xss_clean|required');
+        $this->form_validation->set_rules('cpassword','Confirm Password','trim|xss_clean|matches[password]');
+        if($this->form_validation->run() == FALSE){
+            $this->send_warning(validation_errors());
+        }else{
+            $post_data = $this->array_from_post(array('password'));
+            $this->cm->table_name = "engineers";
+            $this->cm->where = array('ID'=>$ID);
+            if($this->cm->update(array('password'=>$this->hash($post_data['password'])))){
+                $this->send_success('Password Changed Successfully');
+            }else{
+                $this->send_error('Can\'t Change Password Now, Please Try Again');
             }
         }
     }
@@ -198,12 +261,28 @@ class Engineers extends MY_Controller{
     }
 
     public function delete($ID = NULL){
+        $this->cm->table_name = "sites";
+        $this->cm->field_name = "engineerID";
+        $this->cm->primary_key = $ID;
+        $sites = $this->cm->get();
+        if($sites->num_rows() > 0){
+            $this->send_error('Engineer Was Already Added To '.$sites->num_rows().' Site So You Can\'t Delete This Enginner');
+            exit;
+        }
+
         $eng = $this->single_engineer($ID,'array');
         if($this->cm->delete_file('./uploads/'.$eng->photo) && $this->cm->delete_file('./uploads/'.$eng->document)){
             $this->cm->_table_name = "engineers";
             $this->cm->_field_name = "ID";
             $this->cm->_primary_key = $ID;
-            if($this->cm->delete()){
+            $account_delete = $this->cm->delete();
+
+            $this->cm->table_name = "engineersalary";
+            $this->cm->field_name = "engineerID";
+            $this->cm->primary_key = $ID;
+            $salary_delete = $this->cm->delete();
+
+            if($account_delete && $salary_delete){
                 $this->send_success('Engineer Deleted Successfully');
             }else{
                 $this->send_error('Can\'t Delete Right Now, Please Try Again');
